+++ NEWS.md
@@ -5,8 +5,10 @@
 -   Added the `metamean()` function, which can apply several methods to
     estimate study-specific means and their standard errors in order to
     estimate the pooled (difference of) means
--   Added the `metamedian()` function, which is a wrapper for the `qe()`
-    and `pool.med()` functions.
+-   Added the `ob()` function, which applies the method of Ozturk and
+    Balakrishnan (2020) to estimate the pooled median
+-   Added the `metamedian()` function, which is a wrapper for the
+    `qe()`, `ob()`, and `pool.med()` functions.
 
 ### Changes in Version 0.1.6 (2022-06-18)